+++ app/src/main/java/com/example/pesanin/data/motor/Motor.kt
@@ -16,7 +16,7 @@ data class Motor (
     override val quantity: Int,
     override val releaseYear: Date,
     override val color: String,
-    override val price: Int,
+    override val price: Long,
     val engine : String,
     val suspensionType : String,
     val transmisionType : String
